fix(cuda): pinned-host leaks, motion SAD precision, errno fidelity (T-BUGHUNT-CUDA-2026-06-27)
Three CUDA-backend bug classes from the 2026-06-27 bug-hunt sweep, all
fork-local (the Netflix CPU golden gate is CPU-only and untouched).

1. Pinned host-buffer leaks. float_vif_cuda never freed num_host[4]/
   den_host[4]; float_adm_cuda never freed accum_host[FADM_NUM_SCALES].
   Both are page-locked host buffers from vmaf_cuda_buffer_host_alloc that
   were absent from close_fex_cuda and the init free_buffers error path,
   leaking on every vmaf_close(). Added vmaf_cuda_buffer_host_free in both
   close and init-error paths. integer_ms_ssim_cuda was already correct
   (close frees h_ref/h_cmp + the per-scale partial triples) — no change.

2. integer_motion_cuda SAD precision. normalize_and_scale_sad truncated to
   single precision via an intermediate (float) cast and divided by an
   unsigned w*h product, diverging from the CPU twin's
   (double)sad / 256. / (w*h) in integer_motion.c. Recast end-to-end in
   double.

3. speed_temporal/speed_chroma errno squash. Every fail: label reachable
   from CHECK_CUDA_GOTO returned a literal -EIO, discarding the _cuda_err
   the macro had already mapped from the CUresult. Changed all
   macro-reachable fail:/fail_pop:/fail_after_pop: labels to
   return _cuda_err;, matching CHECK_CUDA_RETURN. The two manual
   cuMemcpyDtoH/cuCtxPushCurrent boolean checks keep their literal -EIO.

Reproducer: meson setup core/build-cuda -Denable_cuda=true && ninja -C
core/build-cuda && meson test -C core/build-cuda test_cuda_float_vif_parity
test_cuda_motion3_parity test_cuda_motion_v2_parity
test_cuda_speed_temporal_parity test_cuda_speed_chroma_parity
test_cuda_preallocation_leak  (10/10 pass on RTX 4090, CUDA 13.3).
Pre-existing parity failures in float_adm/cambi/ssimulacra2/float_moment/
psnr_hvs confirmed present on the unmodified baseline (out of scope).

Deliverables: changelog.d/fixed fragment, docs/state.md Recently-closed row,
docs/rebase-notes.md entry, core/src/feature/cuda/AGENTS.md invariants.
No golden assertion modified.

+++ core/src/feature/cuda/integer_motion_cuda.c
@@ -432,7 +432,11 @@ static int append_if_unwritten(VmafFeatureCollector *fc, const char *feature, do
 
 static inline double normalize_and_scale_sad(uint64_t sad, unsigned w, unsigned h)
 {
-    return (float)(sad / 256.) / (w * h);
+    /* Compute end-to-end in double, matching the CPU twin
+     * (integer_motion.c: (double)sad / 256. / (w * h)). A prior
+     * intermediate (float) cast truncated to single precision and the
+     * unsigned (w * h) divisor risked overflow on large frames. */
+    return (double)sad / 256.0 / ((double)w * (double)h);
 }
 
 static int emit_batch_scores(MotionStateCuda *s, VmafFeatureCollector *fc, unsigned batch_start,
